Simplify code reading
